refactor: use `unordered_set` for subscriptions
