+++ src/providers/seventv/SeventvEventApiClient.hpp
@@ -1,22 +1,14 @@
 #pragma once
 
 #include <QString>
-#include <boost/functional/hash.hpp>
 #include <pajlada/signals/signal.hpp>
 #include "providers/seventv/SeventvEventApi.hpp"
 #include "providers/seventv/SeventvEventApiWebsocket.hpp"
 
 #include <atomic>
-#include <vector>
+#include <unordered_set>
 
 namespace chatterino {
-struct SeventvEventApiSubscription {
-    bool operator==(const SeventvEventApiSubscription &rhs) const;
-    bool operator!=(const SeventvEventApiSubscription &rhs) const;
-    QString condition;
-    SeventvEventApiSubscriptionType type;
-};
-
 class SeventvEventApiClient
     : public std::enable_shared_from_this<SeventvEventApiClient>
 {
@@ -42,7 +34,7 @@ public:
 
     bool isSubscribedToEmoteSet(const QString &emoteSetId);
 
-    std::vector<SeventvEventApiSubscription> getSubscriptions() const;
+    std::unordered_set<SeventvEventApiSubscription> getSubscriptions() const;
 
 private:
     void checkHeartbeat();
@@ -50,7 +42,7 @@ private:
 
     eventapi::WebsocketClient &websocketClient_;
     eventapi::WebsocketHandle handle_;
-    std::vector<SeventvEventApiSubscription> subscriptions_;
+    std::unordered_set<SeventvEventApiSubscription> subscriptions_;
 
     std::atomic<std::chrono::time_point<std::chrono::steady_clock>>
         lastHeartbeat_;
@@ -58,17 +50,3 @@ private:
     std::atomic<bool> started_{false};
 };
 }  // namespace chatterino
-
-namespace std {
-template <>
-struct hash<chatterino::SeventvEventApiSubscription> {
-    size_t operator()(const chatterino::SeventvEventApiSubscription &sub) const
-    {
-        std::size_t seed = 0;
-        boost::hash_combine(seed, qHash(sub.condition));
-        boost::hash_combine(seed, sub.type);
-
-        return seed;
-    }
-};
-}  // namespace std
